requirements.txt
pygame==1.9.3
pyscroll==2.17.8
PyTMX==3.21.5
six==1.11.0
tazlib==1.3.2
